Remembered to reimplement the code for scribbling.
I made a really idiotic mistake earlier this afternoon; I copied a
snippet of code for rectifying the offset between the cursor and canvas on
the horizontal axis for the offset on the vertical axis, but clumsily
forgot to change the attribute from left to top..! Ugh!

All the same, I also appended some code for aggregating the scribbling
with the instantiate of the canvas. I am excited about the bits of code
I've managed to compile together! :D